⚙️ Refacto: Destruc and create login component
